Fix: `CustomInput` can only be serialized polymorphically
When trying to upgrade to Kotlin 2.0.0, a compiler error occurs on `CustomInput` due to the applied `CustomInputSerializer` which has a constructor parameter which is not allowed.

This new compilation error is correct. The registered serializer is unusable; `CustomInput` can only be serialized as part of polymorphic serialization. Therefore, a fix is to simply remove this, so such serialization cannot be attempted.