Make `StudyDeployment.deviceDeployed` deterministic
This still relied on a `Clock` which couldn't be injected, thereby creating differing output each time a `StudyDeploymentStatus.Running` would be returned in a test. For this output to work in a `BackwardsCompatibilityTest`, which serializes application service output, this needs to use the time determined by an injected `Clock` which can be made deterministic in tests.